+++ lib/bh/helpers/navbar_helper.rb
@@ -1,11 +1,9 @@
-require 'bh/helpers/base_helper'
 require 'bh/helpers/horizontal_helper'
 require 'bh/helpers/vertical_helper'
 
 module Bh
   # Provides the `navbar` helper.
   module NavbarHelper
-    include BaseHelper
     include HorizontalHelper
     include VerticalHelper
