+++ spec/rails/form/field_helper_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 require 'bh/core_ext/rails/form_for_helper'
-include Bh::FormForHelper
+include Bh::Rails::Helpers
 
 def self.field_helpers_to_test
   types = %w(email number password phone search telephone text url)
